+++ docs/changelog.md
@@ -11,6 +11,12 @@ Here is a highlight of changes in each versions. If you need further details,
 follow [merged Pull request
 pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Amerged).
 
+# UNRELEASED
+
+- Fix inspection of grants on functions. Thanks @dani.
+- Fix conflict between routines and functions ACL.
+
+
 # ldap2pg 6.5.0
 
 - Fix role config not applied to dynamic role rule. Thanks @astentx1.